Website updates and changes. 25 April 2025 Fixed the navigation links for e-waste requests. Added "view.php" for e-waste requests module. Fixed manage users (delete user) styles.
26 April 2025
Major changes made to contents and styles of e-waste requests (view) webpage.
Fixed the manage users (delete user) page title banner.
Fixed the account logout comments not hidden on the webpage.

To-do:
Add a function to display that the account is not registered in the database for login page.
Rework mailing list function.
Possibly dropdown menu for Admin button.
Remember the entries the user typed in the table query forms (Optional).
[Priority] Update Manage E-waste Requests Screening/Acceptance
[Priority] Update Account Registration
[Priority] Update Account Registration Process
[Priority] Update Account Login
[Priority] Update Account Login Process
[To-do] About us webpage.
[To-do] Buy/Sell request webpage.
[To-do] Contact us webpage.
[To-do] Dashboard webpage.
[To-do] Requests history webpage.
[To-do] Services webpage.
[To-do] Tracking webpage.
[To-do] Transactions history webpage.
[To-do] Manage profile webpage.
[To-do] Admin - Statistics webpage.
[Priority] Admin - E-waste requests screen/acceptance webpage.
[Priority] Manage users (Edit).
[Priority] Manage users (Delete confirmation).